+++ src/main/java/org/openpnp/gui/MainFrame.java
@@ -874,10 +874,10 @@ public class MainFrame extends JFrame {
         });
     }
     
-    public void setPlacements(int CompletedPlacemnts, int TotalPlacements, int PlacementsBoard, int TotalPlacementsBoard) {
+    public void setPlacementCompletionStatus(int totalPlacementsCompleted, int totalPlacements, int boardPlacementsCompleted, int boardPlacements) {
         SwingUtilities.invokeLater(() -> {
-        	lblPlacements.setText(" Placements: " + CompletedPlacemnts + " / " + TotalPlacements + " Total | " + PlacementsBoard + " / " + TotalPlacementsBoard + " Selected Board "); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$
-        	prgbrPlacements.setValue((int)(((float)CompletedPlacemnts / (float)TotalPlacements) * 100.0f));
+            lblPlacements.setText(String.format(" Placements: %d / %d Total | %d / %d Selected Board ", totalPlacementsCompleted, totalPlacements, boardPlacementsCompleted, boardPlacements));
+        	prgbrPlacements.setValue((int)(((float)totalPlacementsCompleted / (float)totalPlacements) * 100.0f));
         });
     }