+++ src/main/java/org/openpnp/spi/PnpJobProcessor.java
@@ -1,16 +1,10 @@
 package org.openpnp.spi;
 
-import java.util.List;
-
 import org.openpnp.model.BoardLocation;
 import org.openpnp.model.LengthUnit;
 import org.openpnp.model.Placement;
 
 public interface PnpJobProcessor extends JobProcessor {
-
-    public List<JobPlacement> getJobPlacementsById(String id);
-
-    public List<JobPlacement> getJobPlacementsById(String id, JobPlacement.Status status);
 
     public static class JobPlacement {
         public enum Status {